findAvg {plotRs}	R Documentation

Uses inported Rs data to find the average of each year

Description

Uses inported Rs data to find the average of each year

Usage

findAvg(data, idx)

Arguments

data
list from imported matlab data

idx
the index of the year you like 1 = 2014 2 = 2015 3 = 2016 4 = 2017 5 = 2018

Value

average data

Examples

findAvg(rsIn)

[Package plotRs version 0.1.0 Index]